Update repo URL in README
GitHub Pages has been the primary distribution channel for this chart since March 2024. However, the repo URL in the README has not been updated accordingly.

I would like to decommission the https://helm.twun.io/ repo, but would prefer not to do so without sufficient notice to the community to update their repo URLs.

Updating the README is one necessary step toward this.